Prefix duplicate tab names with folders
when using neovim tabs as oni tabs with "tabs.mode: 'tabs'".

When editing multiple "index.js" files oni will prefix the names of the
non selected tabs with the folder name. This makes it easier to tell
which file is which.

This behaviour differs from "tabs.mode: 'buffers'", where all oni tabs
editing a file with the same name are prefixed. Even the currently
selected one.

Resolves #2420.